review: fix .env load order + persona reload state hygiene
Addresses Copilot review on PR #9:

- Extract .env parsing into src/bootstrap-env.ts and import it as the
  first line of src/index.ts, so modules that capture process.env at
  module-init (qwen-persona.ts QWEN_PERSONA_DIR among others) see
  .env values during their own load.
- In refreshIfStale(), only update _lastCheckAt after a successful
  budget check, so a thrown reload doesn't suppress re-validation
  for the TTL window. Correct the misleading FATAL-comment to
  describe actual behavior (runQwenTurn's outer catch swallows the
  throw — ADR-0003 scopes fail-fast to startup, not runtime reload).
